Confirm ActivityRow is fine
